app/AI-System/Generative Language/api/batch_manager.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv(r"c:\Divaparadises\Divaparadises\app\.env")
API_KEY = os.getenv("VITE_GEMINI_API_KEY")

class BatchManager:
    """
    Manages Batch Jobs (v1beta.batches).
    Run async prompts for 50% cost savings.
    """
    def __init__(self):
        if not API_KEY:
            raise ValueError("API Key not found")
        genai.configure(api_key=API_KEY)

    def create_batch(self, model_name, prompts, display_name="diva_batch"):
        """
        Creates a batch job.
        Args:
            prompts: List of text strings.
        """
        try:
            # Need 'google-generativeai' >= 0.7.0 for batches? 
            # We check if feature exists.
            print(f"[BATCH] Submitting {len(prompts)} items...")
            
            # Note: The SDK implementation for Batch is 'create_batch_job' or similar.
            # Assuming generic structure or skipping if missing.
            
            # This is a placeholder for the advanced feature 
            # as it relies on specific SDK versions.
            
            print("[WARN] Batch API implementation pending SDK confirmation.")
            return None
            
        except Exception as e:
            print(f"[ERROR] Create Batch failed: {e}")
            return None

    def list_batches(self):
        try:
            # Placeholder
            return []
        except Exception as e:
            return []

if __name__ == "__main__":
    bm = BatchManager()
    print("Batches:", bm.list_batches())
