app/AI-System/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google AI Platform
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-auth==2.25.2

# HTTP and File Handling
httpx==0.25.2
aiofiles==23.2.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.25.2
pillow==10.1.0

# Audio/Video Processing
moviepy==1.0.3
pydub==0.25.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0

# Environment and Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
requests==2.31.0
tqdm==4.66.1
colorama==0.4.6